Update project settings
Multi-threaded DLL - Fixes LNK2016 errors and allows compiling without the /force link setting

GenerateDebugInformation: DebugFull - Needed to produce a valid PDB file in release mode

FixedBaseAddress: True - "Tells the operating system to load the program only at its preferred base address. If the preferred base address is unavailable, the operating system does not load the file." (Would we prefer to allow loading the DLL at a different location and then throw if it is wrong, or just fail to load the file altogether?

ImageHasSafeExceptionHandling: Outpost 2 and op2ext do not support this, so seemed proper to disable in NetFixClient as well