Update Log.h to prefer forward declares
Forward declares were added for Windows structs, though it may be
desireable to include a Windows header instead.

The `Packet` class might be better defined as a struct.

The `DumpPacket` method was correct to include the namespace of the
type for its parameter. This was likely not causing a compilation error
before, because the only call to the method had already been commented
out.