Pull game directory from op2ext
This allows attaching a debugger and still finding the correct location of the .ini file. Otherwise if the debugger is attached, the default protocol index of 4 is returned (because .ini file is not found) which makes it appear as if the module is not loading properly.

I wanted to find a solution that would allow Outpost 2 to be buried in a directory deeper than MAX_PATH, but didn't come up with anything useful.