Construct `OPUNetGameSelectWnd` in `DoStart`
This delays construction of the window until the users selects it from
the multiplayer menu.

This avoids early construction of the window, which has constructor
calls which access Outpost2.exe's module address space. Such access
should be delayed until after the module load address check.

In particular, global objects, and components of global objects would be
initialized before the module load address check. Currently
`OPUNetGameProtocol` is constructed as a global object. Hence it should
not contain members, such as `OPUNetGameSelectWnd`, with constructors
that cross the module boundary.