Silence column text warnings in Mingw
The `pszText` field can be used to set information, or to retrieve
information from a list view control. When retrieving information, the
`pszText` field point to a buffer to receive the information, while a
`cchTextMax` field gives the buffer size. As this field is able to point
to a modifiable buffer, it is not declared `const`. As it is not
declared `const`, it generates compiler warnings when a string literal
is assigned to it.